fix(platform): resolve circular dependency between index.ts and paths.ts
Move sortNvmVersionDirs definition to paths.ts and re-export from index.ts.
This breaks the circular dependency where:
- index.ts exported from paths.ts
- paths.ts imported sortNvmVersionDirs from index.ts

Now sortNvmVersionDirs lives in paths.ts alongside findNodeJsDirectories
which uses it, and index.ts re-exports it for backwards compatibility.
